Address #5302
This override was added many years ago to fix an issue (#1747) with the expression that occurred when using variables. This was because Skript could not properly convert variables during parse time. The proper fixes have been made with this branch and so this method is no longer needed. I've updated the test to include code from the original issue.