Implement additional comparator resolution
This enables Skript to compare to objects by converting them to the same type. Currently, this only supports converting Object A to the type of Object B and vice-versa. It is not currently possible for Skript to do Object A to some type C and Object B to some type C. That could be a heavier operation and potentially not work the performance degradation.