make the max target block value 1000
Makes the max target block config value 1000 instead of 100. I'm always needing to change this value, with builders and players having high render distance now with modded clients.

Code is more optimized than it was back in Java 6. The target block now uses ray tracing, and is not looping over blocks in the way it used to.

Cleaned up the description, it doesn't really impact performance, the value just has to exist so it doesn't run away if the player is looking at air.